feat(13-03): delete the whole-tree eager fallback, replace with a hard error
Per D-02 (locked in 13-CONTEXT.md, confirmed by the developer at this plan's
Task 1 checkpoint): the whole-tree eager fallback (ParseAll), its one-shot
latch (eagerFallback) and its exported accessor (FellBackToEager) are
deleted. The scan tier (13-01) and the symbol index (13-02) are the coverage
that makes this safe; an exhausted resolveTiers chain now returns a hard
error wrapping protoregistry.NotFound and naming the unresolved symbol, the
import roots searched, and the symbol index's build/cache state.

The compile-finished log line is replaced in kind, not dropped: it now
reports symbolIndexBuilds/symbolIndexCacheHits/scanResolutions on every
compile instead of the old eagerFallback boolean.

Four tests broken by the deletion are re-pointed at the replacement
mechanism rather than deleted:
- compiler/lib/eager_fallback_visible_test.go -> tier_observability_test.go
- utils/parse_all_deadlock_test.go -> index_build_deadlock_test.go, folded
  with 13-02's already-existing TestIndexBuildRacesParseOne rather than
  keeping both
- utils/lazy_parse_canonical_test.go's afterParseHook now stages the same
  concurrent-insert interleaving directly instead of via ParseAll
- utils/growable_resolver_test.go's two ParseAll-specific tests are deleted
  with their reasoning recorded: the diff-registration behavior they pinned
  lived entirely inside ParseAll's own deleted loop, with no production
  caller of Import/Parse on a lazy registry to re-point at

Also fixes compiler/lib/lazy_load_count_test.go, a caller of the deleted
FellBackToEager accessor the plan's own file list missed (Rule 3: blocking
compile issue).
